feat(category): Add category management

+++ projekt/frontend/src/app/feature/tag/tag-details/tag-details.ts
@@ -5,7 +5,7 @@ import { UUIDTypes } from 'uuid';
 import { CommonModule } from '@angular/common';
 import { CategoryColorService } from '../../category-color.service';
 import { CategoryColor } from '../../../shared/enums/CategoryColor';
-import Tag from '../../../interface/tag/tag';
+import Tag from '../../../interface/tag/Tag';
 
 @Component({
   selector: 'app-tag-details',
@@ -60,7 +60,20 @@ export class TagDetails implements OnInit {
   }
 
   deleteTag(tagId: UUIDTypes): void {
+    if (this.currentIndex === undefined) return;
+    
+    const nextIndex = this.currentIndex < this.tags.length - 1 
+      ? this.currentIndex 
+      : this.currentIndex - 1;
+    
     this.tagsService.deleteTag(tagId);
+    
+    if (nextIndex >= 0 && this.tags.length > 1) {
+      const nextTag = this.tags[nextIndex === this.currentIndex ? nextIndex + 1 : nextIndex];
+      this.router.navigate(['tags', nextTag.id, 'details']);
+    } else {
+      this.router.navigate(['tags']);
+    }
   }
 
   getTagBackgroundColor(categoryColor: CategoryColor): string {
